feat: order

+++ groebner/composeApp/src/commonMain/kotlin/io/printer.kt
@@ -10,7 +10,7 @@ fun printPolynomial(poly: Polynomial): String {
     for (mono in poly.monomials) {
         if (!leadingMono) terms.add(if (mono.coefficient >= 0) "+" else "-")
 
-        if (abs(mono.coefficient) == (1).toLong()) {
+        if (abs(mono.coefficient) == 1L) {
             if (mono.powers.size == 0)
                 terms.add("1")
         } else {
@@ -22,7 +22,7 @@ fun printPolynomial(poly: Polynomial): String {
         }
 
         for (power in mono.powers) {
-            if (power.value == 1) {
+            if (power.value == 1u) {
                 terms.add(power.key.toString())
             } else {
                 terms.add(power.key + "^" + power.value)
